requirements.txt
streamlit~=1.50.0
streamlit-tags
streamlit-calendar

Pillow==9.5.0
pandas~=2.3.3

sqlalchemy~=2.0.43
python-dotenv~=1.1.1

instagrapi~=2.2.1
requests~=2.32.4

pydantic~=2.11.7
openai~=2.2.0
langchain~=0.3.27
langchain-openai~=0.3.35

moviepy==1.0.3
matplotlib
termcolor~=3.1.0

streamlit-autorefresh
heyoo~=0.1.2

openpyxl~=3.1.5
html2text~=2025.4.15msal~=1.31.0
